+++ docs/conf.py
@@ -19,10 +19,14 @@ sys.path.insert(0, os.path.abspath('..'))
 #
 # needs_sphinx = '1.0'
 
+package_root = "."
+
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc',
+extensions = ['sphinxcontrib.extras_require',
+              'sphinx-prompt',
+              'sphinx.ext.autodoc',
               'sphinx.ext.intersphinx',
               'sphinx.ext.todo',
               'sphinx.ext.coverage',
@@ -68,7 +72,7 @@ release = '{:s}-alpha'.format(version)  # Include alpha/beta/rc tags.
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = None
+language = 'en'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -92,7 +96,7 @@ html_theme = 'sphinx_rtd_theme'
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-html_logo = os.path.join(os.path.abspath('.'), 'figures', 'pysatnasa_logo.jpg')
+html_logo = os.path.join(os.path.abspath('.'), 'figures', 'logo.png')
 html_theme_options = {'logo_only': True}
 
 # Add any paths that contain custom static files (such as style sheets) here,