+++ pysatNASA/instruments/cnofs_vefi.py
@@ -6,7 +6,7 @@ onboard the Communication and Navigation Outage Forecasting
 System (C/NOFS) satellite. Downloads data from the
 NASA Coordinated Data Analysis Web (CDAWeb).
 
-Description from CDAWeb:
+Description from CDAWeb (with updated link):
 
 The DC vector magnetometer on the CNOFS spacecraft is a three axis, fluxgate
 sensor with active thermal control situated on a 0.6m boom.  This magnetometer
@@ -15,7 +15,7 @@ sec with a range of .. 45,000 nT.  Its primary objective on the CNOFS
 spacecraft is to enable an accurate V x B measurement along the spacecraft
 trajectory.  In order to provide an in-flight calibration of the magnetic field
 data, we compare the most recent POMME model (the POtsdam Magnetic Model of the
-Earth, http://geomag.org/models/pomme5.html) with the actual magnetometer
+Earth, https://geomag.us/models/pomme5.html) with the actual magnetometer
 measurements to help determine a set of calibration parameters for the gains,
 offsets, and non-orthogonality matrix of the sensor axes.  The calibrated
 magnetic field measurements are provided in the data file here. The VEFI
@@ -63,10 +63,10 @@ import functools
 import numpy as np
 
 from pysat.instruments.methods import general as mm_gen
-from pysat import logger
 
 from pysatNASA.instruments.methods import cdaweb as cdw
 from pysatNASA.instruments.methods import cnofs as mm_cnofs
+from pysatNASA.instruments.methods import general as mm_nasa
 
 # ----------------------------------------------------------------------------
 # Instrument attributes
@@ -85,18 +85,8 @@ _test_dates = {'': {'dc_b': dt.datetime(2009, 1, 1)}}
 # Instrument methods
 
 
-def init(self):
-    """Initialize the Instrument object with instrument specific values.
-
-    Runs once upon instantiation.
-
-    """
-    logger.info(mm_cnofs.ackn_str)
-    self.acknowledgements = mm_cnofs.ackn_str
-    self.references = '\n'.join((mm_cnofs.refs['mission'],
-                                 mm_cnofs.refs['vefi']))
-
-    return
+# Use standard init routine
+init = functools.partial(mm_nasa.init, module=mm_cnofs, name=name)
 
 
 def clean(self):